Use BatchCreateSessionsRequest to create sessions. (googleapis/google-cloud-cpp-spanner#924)
* Use BatchCreateSessionsRequest to create sessions.

Part of googleapis/google-cloud-cpp-spanner#307

* fix return type to work with all compilers.

* fix windows by avoiding narrowing.

* Fix signed/unsigned and narrowing issues.

* Fix the new tests to use BatchCreateSessions

Also fixed a typo in a comment/string

+++ google/cloud/spanner/internal/connection_impl.h
@@ -145,7 +145,7 @@ class ConnectionImpl : public Connection,
 
   // `SessionManager` methods; used by the `SessionPool`
   StatusOr<std::vector<std::unique_ptr<Session>>> CreateSessions(
-      size_t num_sessions) override;
+      int num_sessions) override;
 
   Database db_;
   std::shared_ptr<SpannerStub> stub_;
